+++ code/xcvm/cosmwasm/contracts/gateway/src/lib.rs
@@ -8,3 +8,4 @@ pub mod exec;
 pub mod state;
 mod events;
 mod topology;
+mod prelude;
