+++ code/xcvm/cosmwasm/contracts/gateway/src/prelude.rs
@@ -1,0 +1,3 @@
+pub use cosmwasm_std::{
+	to_binary, wasm_execute, Binary, Coin,  Addr,
+};
